requirements.txt
aiofile==3.9.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiopath==0.6.11
aiosignal==1.3.1
aiosmtplib==3.0.2
alembic==1.14.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
backports.tarfile==1.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.2
caio==0.9.17
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
constantly==23.10.4
coverage==7.7.0
crashtest==0.4.1
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.4
dnspython==2.7.0
dulwich==0.22.7
ecdsa==0.19.0
email_validator==2.2.0
Faker==33.1.0
fastapi==0.115.8
fastapi-cli==0.0.7
fastapi-limiter==0.1.6
fastapi-mail==1.4.2
fastapi-users==14.0.1
fastjsonschema==2.21.1
filelock==3.16.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
iniconfig==2.0.0
installer==0.7.0
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.5
jmespath==1.0.1
jose==1.0.0
keyring==25.6.0
libgravatar==1.0.4
lxml==5.3.0
makefun==1.15.6
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
names==0.3.0
orjson==3.10.15
packaging==24.2
parsel==1.9.1
passlib==1.7.4
pika==1.3.2
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.0
poetry-core==2.0.0
propcache==0.2.0
Protego==0.3.1
psycopg2-binary==2.9.10
pwdlib==0.2.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.10.1
pyOpenSSL==24.3.0
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
queuelib==1.7.0
RapidFuzz==3.11.0
redis==5.2.1
redis-lru==0.1.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
Scrapy==2.12.0
SecretStorage==3.3.3
service-identity==24.2.0
setuptools==77.0.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.45.3
tldextract==5.1.3
tomlkit==0.13.2
trove-classifiers==2025.1.10.15
Twisted==24.11.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
unicorn==2.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.1
w3lib==2.2.1
watchfiles==1.0.4
websockets==14.1
yarl==1.17.2
zipp==3.21.0
zope.interface==7.2
